refectoring emd_1d and variants
