+++ test/test_sliced.py
@@ -10,11 +10,6 @@ import pytest
 
 import ot
 from ot.sliced import get_random_projections
-from ot.sliced import emd1D
-
-from ot.backend import get_backend_list
-
-backend_list = get_backend_list()
 
 
 def test_get_random_projections():
@@ -89,35 +84,3 @@ def test_1d_sliced_equals_emd():
     res = ot.sliced_wasserstein_distance(x, y, a, u, 10, seed=42)
     expected = ot.emd2_1d(x.squeeze(), y.squeeze(), a, u)
     np.testing.assert_almost_equal(res ** 2, expected)
-
-
-@pytest.mark.parametrize('nx', backend_list)
-def test_emd1D(nx):
-    from scipy.stats import wasserstein_distance
-
-    rng = np.random.RandomState(0)
-
-    n = 100
-    x = np.linspace(0, 5, n)
-    rho_u = np.abs(rng.randn(n))
-    rho_u /= rho_u.sum()
-    rho_v = np.abs(rng.randn(n))
-    rho_v /= rho_v.sum()
-
-    xb = nx.from_numpy(x)
-    rho_ub = nx.from_numpy(rho_u)
-    rho_vb = nx.from_numpy(rho_v)
-
-    # test 1 : emd1D should be close to scipy W_1 implementation
-    np.testing.assert_almost_equal(emd1D(xb, xb, rho_ub, rho_vb, p=1),
-                                   wasserstein_distance(x, x, rho_u, rho_v))
-
-    # test 2 : emd1D should be close to one when only translating the support
-    np.testing.assert_almost_equal(emd1D(xb, xb + 1, p=2),
-                                   1.)
-
-    # test 3 : arrays test
-    X = np.stack((np.linspace(0, 5, n), np.linspace(0, 5, n) * 10), -1)
-    Xb = nx.from_numpy(X)
-    res = emd1D(Xb, Xb, rho_ub, rho_vb, p=2)
-    np.testing.assert_almost_equal(100 * res[0], res[1], decimal=4)
